build: split examples into a separate goal
Some examples are epend on MWare, which we do not need. The issue
is that the MWare has a compilation issue to to stdint.h not
being included. In ENV.h, this header seems to be included
specifically for TIVAWARE and nto for MWARE (unclear why). With
a separate goal, we can build all the other packages (for TivaWare),
and leave the examples alone.